backend/requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

matplotlib
pandas
numpy
scikit-learn
scipy
sqlalchemy
pillow
tensorflow
keras
streamlit
streamlit-folium
streamlit_lottie
altair
pillow